bugfixes and version update

+++ h5rdmtoolbox/__init__.py
@@ -91,4 +91,5 @@ def clean_temp_data():
                     for fd in failed_dirs:
                         f.writelines(f'{fd}\n')
             else:
-                failed_dirs_file.unlink()
+                if failed_dirs_file.exists():
+                    failed_dirs_file.unlink()
